nested class player
moved player to be a nested class within match
